Move from `/details` to `/src` (on_audio_query)
