+++ packages/on_audio_query_platform_interface/lib/details/extensions/models_extensions.dart
@@ -1,4 +1,4 @@
-part of on_audio_query_helper;
+part of on_audio_query_core;
 
 /// Used to convert a [List<dynamic>] into a Model.
 ///
