+++ packages/on_audio_query_platform_interface/lib/details/log/log_config.dart
@@ -1,4 +1,4 @@
-part of on_audio_query_helper;
+part of on_audio_query_core;
 
 /// Used to configure the logging behavior.
 class LogConfig {
