+++ packages/on_audio_query_platform_interface/lib/details/models/genre_model.dart
@@ -1,4 +1,4 @@
-part of on_audio_query_helper;
+part of on_audio_query_core;
 
 /// [GenreModel] that contains all [Genre] Information.
 class GenreModel {
