+++ packages/on_audio_query_platform_interface/lib/details/types/audios_from_type.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: constant_identifier_names
 
-part of on_audio_query_helper;
+part of on_audio_query_core;
 
 /// Defines where audios will be acquired.
 enum AudiosFromType {
