+++ packages/on_audio_query_platform_interface/lib/details/types/sort_types/playlist_sort_type.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: constant_identifier_names
 
-part of on_audio_query_helper;
+part of on_audio_query_core;
 
 /// Defines the sort type used for [queryPlaylists].
 enum PlaylistSortType {
