+++ packages/on_audio_query_platform_interface/lib/details/types/with_filters_type.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: constant_identifier_names
 
-part of on_audio_query_helper;
+part of on_audio_query_core;
 
 /// Defines the type of Filter.
 /// Each type has a subtype, [10] in total.
