+++ packages/on_audio_query_platform_interface/lib/method_channel_on_audio_query.dart
@@ -19,7 +19,7 @@ import 'package:flutter/services.dart';
 
 import 'on_audio_query_platform_interface.dart';
 
-import 'package:on_audio_query_platform_interface/details/on_audio_query_helper.dart';
+import 'package:on_audio_query_platform_interface/details/on_audio_query_core.dart';
 
 const MethodChannel _channel = MethodChannel('com.lucasjosino.on_audio_query');
